+++ test/src/common/glyph/generic_glyph_base_test.dart
@@ -162,7 +162,8 @@ void main() {
     });
 
     test('stores a base64 preview that decodes to the input SVG', () {
-      const svg = '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">'
+      const svg =
+          '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">'
           '<path d="M0 0h24v24H0z"/></svg>';
       final glyph = GenericGlyph.fromSvg('box', svg);
       expect(utf8.decode(base64Decode(glyph.metadata.preview!)), svg);